web_frontend/requirements.txt
# Django Web Frontend Requirements
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.6
